update go.mod for clusteragent endpoint
update go mod

linter fix

move getclusteragentendpoint to config utils
