+++ comp/forwarder/defaultforwarder/default_forwarder.go
@@ -26,7 +26,6 @@ import (
 	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
 	"github.com/DataDog/datadog-agent/pkg/config/utils"
 	"github.com/DataDog/datadog-agent/pkg/telemetry"
-	"github.com/DataDog/datadog-agent/pkg/util/clusteragent"
 	"github.com/DataDog/datadog-agent/pkg/util/filesystem"
 	"github.com/DataDog/datadog-agent/pkg/version"
 )
@@ -185,7 +184,7 @@ func NewOptionsWithResolvers(config config.Component, log log.Component, domainR
 
 	// domainforwarder to local DCA for autoscaling failover metrics
 	if config.GetBool("autoscaling.failover.enabled") && config.GetBool("cluster_agent.enabled") {
-		if domain, err := clusteragent.GetClusterAgentEndpoint(); err != nil {
+		if domain, err := utils.GetClusterAgentEndpoint(); err != nil {
 			log.Errorf("Could not get cluster agent endpoint for autoscaling failover metrics: %s", err)
 		} else if authToken, err := security.GetClusterAgentAuthToken(config); err != nil {
 			log.Errorf("Failed to get cluster agent auth token: ", err)
