+++ pkg/util/clusteragent/clusteragent_test.go
@@ -279,11 +279,8 @@ type clusterAgentSuite struct {
 }
 
 const (
-	clusterAgentServiceName = "DATADOG_CLUSTER_AGENT"
-	clusterAgentServiceHost = clusterAgentServiceName + "_SERVICE_HOST"
-	clusterAgentServicePort = clusterAgentServiceName + "_SERVICE_PORT"
-	clusterAgentTokenValue  = "01234567890123456789012345678901"
-	clcRunnerIP             = "10.92.1.39"
+	clusterAgentTokenValue = "01234567890123456789012345678901"
+	clcRunnerIP            = "10.92.1.39"
 )
 
 func (suite *clusterAgentSuite) SetupTest() {
@@ -292,16 +289,6 @@ func (suite *clusterAgentSuite) SetupTest() {
 	suite.config.SetWithoutSource("cluster_agent.url", "")
 	suite.config.SetWithoutSource("cluster_agent.kubernetes_service_name", "")
 	suite.config.SetWithoutSource("clc_runner_host", clcRunnerIP)
-	os.Unsetenv(clusterAgentServiceHost)
-	os.Unsetenv(clusterAgentServicePort)
-}
-
-func (suite *clusterAgentSuite) TestGetClusterAgentEndpointEmpty() {
-	suite.config.SetWithoutSource("cluster_agent.url", "")
-	suite.config.SetWithoutSource("cluster_agent.kubernetes_service_name", "")
-
-	_, err := GetClusterAgentEndpoint()
-	require.NotNil(suite.T(), err)
 }
 
 func (suite *clusterAgentSuite) TestGetClusterAgentAuthTokenEmpty() {
@@ -358,64 +345,6 @@ func (suite *clusterAgentSuite) TestGetClusterAgentAuthTokenTooShort() {
 	require.Nil(suite.T(), err, fmt.Sprintf("%v", err))
 
 	_, err = security.GetClusterAgentAuthToken(suite.config)
-	require.NotNil(suite.T(), err, fmt.Sprintf("%v", err))
-}
-
-func (suite *clusterAgentSuite) TestGetClusterAgentEndpointFromUrl() {
-	suite.config.SetWithoutSource("cluster_agent.url", "https://127.0.0.1:8080")
-	suite.config.SetWithoutSource("cluster_agent.kubernetes_service_name", "")
-	_, err := GetClusterAgentEndpoint()
-	require.Nil(suite.T(), err, fmt.Sprintf("%v", err))
-
-	suite.config.SetWithoutSource("cluster_agent.url", "https://127.0.0.1")
-	_, err = GetClusterAgentEndpoint()
-	require.Nil(suite.T(), err, fmt.Sprintf("%v", err))
-
-	suite.config.SetWithoutSource("cluster_agent.url", "127.0.0.1")
-	endpoint, err := GetClusterAgentEndpoint()
-	require.Nil(suite.T(), err, fmt.Sprintf("%v", err))
-	assert.Equal(suite.T(), "https://127.0.0.1", endpoint)
-
-	suite.config.SetWithoutSource("cluster_agent.url", "127.0.0.1:1234")
-	endpoint, err = GetClusterAgentEndpoint()
-	require.Nil(suite.T(), err, fmt.Sprintf("%v", err))
-	assert.Equal(suite.T(), "https://127.0.0.1:1234", endpoint)
-}
-
-func (suite *clusterAgentSuite) TestGetClusterAgentEndpointFromUrlInvalid() {
-	suite.config.SetWithoutSource("cluster_agent.url", "http://127.0.0.1:8080")
-	suite.config.SetWithoutSource("cluster_agent.kubernetes_service_name", "")
-	_, err := GetClusterAgentEndpoint()
-	require.NotNil(suite.T(), err)
-
-	suite.config.SetWithoutSource("cluster_agent.url", "tcp://127.0.0.1:8080")
-	_, err = GetClusterAgentEndpoint()
-	require.NotNil(suite.T(), err)
-}
-
-func (suite *clusterAgentSuite) TestGetClusterAgentEndpointFromKubernetesSvc() {
-	suite.config.SetWithoutSource("cluster_agent.url", "")
-	suite.config.SetWithoutSource("cluster_agent.kubernetes_service_name", "datadog-cluster-agent")
-	suite.T().Setenv(clusterAgentServiceHost, "127.0.0.1")
-	suite.T().Setenv(clusterAgentServicePort, "443")
-
-	endpoint, err := GetClusterAgentEndpoint()
-	require.Nil(suite.T(), err, fmt.Sprintf("%v", err))
-	assert.Equal(suite.T(), "https://127.0.0.1:443", endpoint)
-}
-
-func (suite *clusterAgentSuite) TestGetClusterAgentEndpointFromKubernetesSvcEmpty() {
-	suite.config.SetWithoutSource("cluster_agent.url", "")
-	suite.config.SetWithoutSource("cluster_agent.kubernetes_service_name", "datadog-cluster-agent")
-	suite.T().Setenv(clusterAgentServiceHost, "127.0.0.1")
-	suite.T().Setenv(clusterAgentServicePort, "")
-
-	_, err := GetClusterAgentEndpoint()
-	require.NotNil(suite.T(), err, fmt.Sprintf("%v", err))
-
-	suite.T().Setenv(clusterAgentServiceHost, "")
-	suite.T().Setenv(clusterAgentServicePort, "443")
-	_, err = GetClusterAgentEndpoint()
 	require.NotNil(suite.T(), err, fmt.Sprintf("%v", err))
 }
